+++ doc/releasenotes.rst
@@ -12,7 +12,7 @@ Release Notes
 ------------------
 
 * Extend visualization routines: Add histogram, sample number, epsilon
-  trajectory, and credible interval plots.
+  trajectory, model probability, and credible interval plots.
 * Test visualization routines on travis.